+++ src/modules/auth/services/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { KakaoUserAfterAuth } from 'src/common/decorators/user.decorator';
-import { UsersService } from '../users/users.service';
+import { UsersService } from '../../users/services/users.service';
 import { JwtService } from '@nestjs/jwt';
 import { UserInvariantViolationException } from 'src/common/exceptions/custom.errors';
 import { WsException } from '@nestjs/websockets';
