+++ src/modules/comments/cocomments/cocomments.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule } from '@nestjs/config';
-import { Cocomment } from './cocomments.entity';
+import { Cocomment } from './entities/cocomments.entity';
 import { CocommentsController } from './cocomments.controller';
-import { CocommentsService } from './cocomments.service';
+import { CocommentsService } from './services/cocomments.service';
 @Module({
     imports: [TypeOrmModule.forFeature([Cocomment]), ConfigModule],
     controllers: [CocommentsController],
