+++ src/modules/comments/comments.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule } from '@nestjs/config';
-import { Comment } from './comments.entity';
+import { Comment } from './entities/comments.entity';
 import { CommentsController } from './comments.controller';
-import { CommentsService } from './comments.service';
-import { Cocomment } from './cocomments/cocomments.entity';
+import { CommentsService } from './services/comments.service';
+import { Cocomment } from './cocomments/entities/cocomments.entity';
 
 @Module({
     imports: [TypeOrmModule.forFeature([Comment, Cocomment]), ConfigModule],
