+++ src/modules/comments/entities/comments.entity.ts
@@ -1,8 +1,8 @@
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne, OneToMany } from 'typeorm';
-import { User } from '../users/entities/users.entity';
-import { Task } from '../tasks/entities/tasks.entity';
+import { User } from '../../users/entities/users.entity';
+import { Task } from '../../tasks/entities/tasks.entity';
 import { BaseEntity } from 'src/common/entities/base.entity';
-import { Cocomment } from './cocomments/cocomments.entity';
+import { Cocomment } from '../cocomments/entities/cocomments.entity';
 
 @Entity()
 export class Comment extends BaseEntity {
