+++ src/modules/master-portfolios/master-portfolios.controller.ts
@@ -15,7 +15,7 @@ import { ConfigService } from '@nestjs/config';
 import { ApiBody, ApiOperation, ApiParam, ApiTags } from '@nestjs/swagger';
 import { PaginatedResponseDto } from 'src/common/response/paginated-response.dto';
 import { UserMasterPortfoliosResponseDto } from './dtos/user-master-portfolios-response.dto';
-import { MasterPortfoliosService } from './master-portfolios.service';
+import { MasterPortfoliosService } from './services/master-portfolios.service';
 import { User } from 'src/common/decorators/user.decorator';
 import { MasterPortfolioRequestDto } from './dtos/master-portfolio-request.dto';
 import {
