+++ src/modules/portfolio-corrections/portfolio-corrections.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PortfolioCorrectionsService } from './portfolio-corrections.service';
+import { PortfolioCorrectionsService } from './services/portfolio-corrections.service';
 import { PortfolioCorrectionsController } from './portfolio-corrections.controller';
 import { LLMModule } from 'src/infra/llm/llm.module';
 import { PromptLoader } from 'src/common/utils/prompt.loader';
