+++ src/modules/portfolio-corrections/services/portfolio-corrections.service.ts
@@ -3,18 +3,18 @@ import { PromptLoader } from 'src/common/utils/prompt.loader';
 import { LLMService } from 'src/infra/llm/llm.service';
 import { correctionSchema, Correction } from 'src/infra/llm/schemas/portfolio-correction.schema';
 import z from 'zod';
-import { PortfolioCorrection } from './entities/portfolio-correction.entity';
+import { PortfolioCorrection } from '../entities/portfolio-correction.entity';
 import { QueryRunner, Repository } from 'typeorm';
-import { AICorrection } from './entities/ai-correction.entity';
+import { AICorrection } from '../entities/ai-correction.entity';
 import { PaginatedResponseDto } from 'src/common/response/paginated-response.dto';
-import { UserPortfolioCorrectionResponseDto } from './dtos/user-portfolio-correction-response.dto';
+import { UserPortfolioCorrectionResponseDto } from '../dtos/user-portfolio-correction-response.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AIGenerationAlreadyExists } from 'src/common/exceptions/custom.errors';
-import { Project } from '../projects/entities/projects.entity';
-import { CreatePortfolioCorrectionDto } from './dtos/create-corrections.dto';
-import { RAGData } from './entities/rag-data.entity';
+import { Project } from '../../projects/entities/projects.entity';
+import { CreatePortfolioCorrectionDto } from '../dtos/create-corrections.dto';
+import { RAGData } from '../entities/rag-data.entity';
 import { RAGDataType } from 'src/common/enums/rag-data-type.enum';
-import { ProjectResponseDto } from './dtos/project-response.dto';
+import { ProjectResponseDto } from '../dtos/project-response.dto';
 import { PortfolioCorrectionStatus } from 'src/common/enums/portfolio-correction-status.enum';
 
 async function checkCorrectionExists(qr: QueryRunner, correctionId: number) {
