+++ src/modules/projects/projects.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ProjectsController } from './projects.controller';
-import { ProjectsService } from './projects.service';
+import { ProjectsService } from './services/projects.service';
 import { Project } from './entities/projects.entity';
 import { UserProject } from '../mappings/user-projects/userProjects.entity';
 import { RedisModule } from '../../infra/redis/redis.module';
