+++ src/modules/tasks/tasks.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { TasksController } from './tasks.controller';
-import { TasksService } from './tasks.service';
+import { TasksService } from './services/tasks.service';
 import { Task } from './entities/tasks.entity';
 import { Step } from '../steps/entities/steps.entity';
 import { UserProject } from '../mappings/user-projects/userProjects.entity';
@@ -10,7 +10,7 @@ import { Manager } from '../mappings/managers/managers.entity';
 import { TaskFile } from '../mappings/task-files/task-files.entity';
 import { UploadService } from '../../infra/upload/upload.service';
 import { Project } from '../projects/entities/projects.entity';
-import { Comment } from '../comments/comments.entity';
+import { Comment } from '../comments/entities/comments.entity';
 import { UsersModule } from '../users/users.module';
 
 @Module({
